Compatibility changes

+++ Controllers/RoomController.php
@@ -1,7 +1,7 @@
 <?php
 
     namespace Controllers;
-    use DAO\RoomDAO as RoomDAO;
+    use DAO\JRoomDAO as RoomDAO;
     use Models\Room as Room;
 
     class RoomController {
@@ -13,7 +13,7 @@
             $this-> roomDAO = new RoomDAO();
         }
 
-        function addRoom($cinemaID, $name, $capacity) {
+        function addRoom($cinemaID, $name, $capacity, $ticketValue) {
             if ($this-> validateRoomName($name, $cinemaID)){
                 $this-> showAddView($cinemaID, "Nombre de sala ya existente en el cine elegido. Intente con otro.", 0);
             } else {
@@ -22,7 +22,12 @@
                 $room-> setCinemaID($cinemaID);
                 if ($capacity >0){
                     $room-> setCapacity($capacity);
-                    $this-> roomDAO-> Add($room);
+                    if ($ticketValue > 0){
+                        $room-> setTicketValue($ticketValue);
+                        $this-> roomDAO-> Create($room);
+                    } else {
+                        $this-> showAddView("Intente usar un precio mayor a 0.");
+                    }
                     $this-> showAddView($cinemaID, "Sala agregada con éxito.", 1);
                 } else {
                     $this-> showAddView($cinemaID, "Intente usar una capacidad mayor a 0.", 0);
@@ -32,14 +37,14 @@
 
         function removeRoom($id){
             $this-> roomDAO-> Delete($id);
-            $roomsList = $this-> roomDAO-> GetAll();
+            $roomsList = $this-> roomDAO-> ReadAll();
             $message = "Sala eliminada con éxito.";
             require_once(VIEWS_PATH."index.php");
         }
 
-        function editRoom ($id, $name, $capacity){
+        function editRoom ($id, $name, $capacity, $ticketValue){
             $room = new Room();
-            $room = $this-> roomDAO-> GetByID($id);
+            $room = $this-> roomDAO-> ReadByID($id);
             $cinemaID = $room-> getCinemaID();
             if ($this-> validateRoomEdit($id, $name, $cinemaID)){
                 $this-> showEditView($id, "Nombre de sala ya existente en el cine elegido. Intente con otro.", 0);
@@ -47,11 +52,16 @@
                 if ($capacity>=0) {
                     $room = new Room();
                     $room-> setName($name);
-                    $room-> setCapacity($capacity);
-                    $room-> setCinemaID($cinemaID);
-                    $this-> roomDAO-> Add($room);
-                    $this-> roomDAO-> Delete($id);
-                    $this-> showEditView($room-> getID(), "Sala editada con éxito.", 1);
+                    if ($ticketValue>=0) {
+                        $room-> setTicketValue($ticketValue);
+                        $room-> setCapacity($capacity);
+                        $room-> setCinemaID($cinemaID);
+                        $this-> roomDAO-> Create($room);
+                        $this-> roomDAO-> Delete($id);
+                        $this-> showEditView($room-> getID(), "Sala editada con éxito.", 1);
+                    } else {
+                        $this-> showEditView($id, "Intente usar un valor de entrada positivo.");
+                    }
                 } else {
                     $this-> showEditView($id, "Intente usar una capacidad mayor a 0.", 0);
                 }
@@ -69,12 +79,12 @@
         }
 
         function showEditView ($id, $message = "", $messageCode = 0){
-            $room = $this-> roomDAO-> getByID($id);
+            $room = $this-> roomDAO-> ReadByID($id);
             require_once(VIEWS_PATH."room-edit.php");
         }
         
         private function validateRoomName($name, $cinemaID){
-            $roomList = $this-> roomDAO-> getAll();
+            $roomList = $this-> roomDAO-> ReadAll();
             foreach ($roomList as $room){
                 if ($room-> getCinemaID() == $cinemaID && $room-> getName() === $name){
                     return true;
@@ -85,7 +95,7 @@
         
         private function validateRoomEdit($id, $name, $cinemaID){
             $room = new Room();
-            $room = $this-> roomDAO-> GetByID($id);
+            $room = $this-> roomDAO-> ReadByID($id);
             if ($room-> getName() != $name){
                 return $this-> validateRoomName($name, $cinemaID);
             }
